Project structure overhaul
Instead of having the popup be built by Vite and the extension itself be
built by extension-cli, this commit moves the popup code into the root
folder and creates individual Vite configs for the different parts of
the extension.

This way, we have more control over the project build process, which in
turn allows us to easily build for either Firefox or Chromium depending
on the chosen build script.